feat: add mobile style
