+++ frontend/src/components/EnterChoice/EnterChoice.scss
@@ -63,3 +63,9 @@
 .app-enter-buttons__button:focus-visible {
 	box-shadow: 0 0 0 3px #fd7f34;
 }
+
+@media (max-width: 850px) {
+	.app-enter-buttons__button {
+		font-size: 1.2rem;
+	}
+}
