+++ frontend/src/components/Poker-table/Poker_table.scss
@@ -38,12 +38,12 @@
   top: 33%;
   left: 37%;
   display: flex;
-  width: 25%;
+  width: 30%;
   height: 10%;
   border: 1px solid #d5e3c3;
   border-radius: 8px;
-  min-height: 80px;
-  min-width: 270px;
+  // min-height: 80px;
+  // min-width: 270px;
 }
 
 .bank__container {
